[Robot Model] refactor aerial robot model to support general multirotor as the base robot model
