+++ aerial_robot_model/CMakeLists.txt
@@ -52,7 +52,7 @@ generate_messages(
 
 catkin_package(
   INCLUDE_DIRS include test
-  LIBRARIES transformable_aerial_robot_model transformable_aerial_robot_model_ros numerical_jacobians
+  LIBRARIES aerial_robot_model aerial_robot_model_ros numerical_jacobians robot_model_pluginlib
   CATKIN_DEPENDS eigen_conversions geometry_msgs interactive_markers kalman_filter kdl_parser message_runtime pluginlib sensor_msgs spinal std_msgs tf tf_conversions tf2_eigen tf2_geometry_msgs tf2_kdl tf2_ros tf_conversions urdf visualization_msgs
   DEPENDS orocos_kdl urdfdom_headers
 )
@@ -73,17 +73,23 @@ include_directories(
   ${EIGEN3_INCLUDE_DIRS}
 )
 
-add_library(transformable_aerial_robot_model
-  src/transformable_aerial_robot_model/robot_model.cpp
-  src/transformable_aerial_robot_model/jacobians.cpp
-  src/transformable_aerial_robot_model/kinematics.cpp
-  src/transformable_aerial_robot_model/stability.cpp
-  src/transformable_aerial_robot_model/statics.cpp)
-target_link_libraries(transformable_aerial_robot_model ${catkin_LIBRARIES} ${orocos_kdl_LIBRARIES} ${EIGEN3_LIBRARIES})
 
-add_library(transformable_aerial_robot_model_ros src/transformable_aerial_robot_model/robot_model_ros.cpp)
-target_link_libraries(transformable_aerial_robot_model_ros transformable_aerial_robot_model ${catkin_LIBRARIES} ${orocos_kdl_LIBRARIES} ${EIGEN3_LIBRARIES})
-add_dependencies(transformable_aerial_robot_model_ros ${PROJECT_NAME}_generate_messages_cpp spinal_generate_messages_cpp)
+add_library(aerial_robot_model
+  src/base_model/robot_model.cpp
+  src/transformable_model/robot_model.cpp
+  src/transformable_model/jacobians.cpp
+  src/transformable_model/kinematics.cpp
+  src/transformable_model/stability.cpp
+  src/transformable_model/statics.cpp)
+target_link_libraries(aerial_robot_model ${catkin_LIBRARIES} ${orocos_kdl_LIBRARIES} ${EIGEN3_LIBRARIES})
+
+add_library(aerial_robot_model_ros src/base_model/robot_model_ros.cpp)
+target_link_libraries(aerial_robot_model_ros aerial_robot_model ${catkin_LIBRARIES} ${orocos_kdl_LIBRARIES} ${EIGEN3_LIBRARIES})
+add_dependencies(aerial_robot_model_ros ${PROJECT_NAME}_generate_messages_cpp spinal_generate_messages_cpp)
+
+add_library(robot_model_pluginlib src/plugin/multirotor_robot_model.cpp)
+target_link_libraries(robot_model_pluginlib ${catkin_LIBRARIES} aerial_robot_model)
+
 
 add_library(servo_bridge src/servo_bridge/servo_bridge.cpp)
 target_link_libraries(servo_bridge ${catkin_LIBRARIES})
@@ -98,7 +104,7 @@ target_link_libraries(interactive_marker_tf_broadcaster ${catkin_LIBRARIES})
 
 ## for rostest
 add_library(numerical_jacobians test/aerial_robot_model/numerical_jacobians.cpp)
-target_link_libraries(numerical_jacobians transformable_aerial_robot_model ${catkin_LIBRARIES})
+target_link_libraries(numerical_jacobians aerial_robot_model ${catkin_LIBRARIES})
 
 
 install(DIRECTORY include/${PROJECT_NAME}/ test/${PROJECT_NAME}/
