+++ robots/hydrus/include/hydrus/hydrus_robot_model.h
@@ -37,7 +37,7 @@
 
 #include <aerial_robot_model/transformable_aerial_robot_model.h>
 
-class HydrusRobotModel : public aerial_robot_model::RobotModel {
+class HydrusRobotModel : public aerial_robot_model::transformable::RobotModel {
 public:
   HydrusRobotModel(bool init_with_rosparam = true,
                    bool verbose = false,
